pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-infra-lab"
version = "0.1.0"
description = "Small repo for full-weight coding post-training and infra evaluation."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "datasets>=2.20.0",
  "evaluate>=0.4.2",
  "huggingface-hub>=0.24.0",
  "orjson>=3.10.0",
  "pyyaml>=6.0.1",
  "torch>=2.4.0",
  "transformers>=4.44.0",
  "accelerate>=0.34.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

